Reducing code repetition and making code more readable
